Fix: 배치 작업 수정 #35
- Synchronized 관련
: 지금 Queue가 어플리케이션 전체에 딱 하나만 생성되게 되어 있는데 (= A사용자랑 B사용자가 프로필을 바꾸면 같은 하나의 Queue에 이전 프로필 사진이 저장되는 구조) 그럼 그래도 멀티스레딩을 좀 고려해야하지 않나 싶은 생각 → 근데 Synchronize는 성능을 좀 떨어뜨리는 측면이 있다 해서 concurrentlinkedQueue를 사용하면 데이터 일관성 유지하면서도 오버헤드가 아니지 않을까 싶은 생각으로 도입해봤어요

- Exception 처리는 S3 Service의 deleteFileFromS3에 IMAGENOTFOUNTEXCEPTION이 선언 안되어 있어서 (선언하면 여기저기 exception을 받아줘야해서..) errorcode로 우선 분류했는데 뭐가 더 좋을까요
- 알람 배치는.. 좀 더 알아볼게요 한 번에 대박적이게 날리고 싶음
